requirements.txt
# LangGraph + LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# Observability
langfuse>=2.0.0

# Article Extraction
trafilatura>=1.12.0
beautifulsoup4>=4.12.0
requests>=2.31.0
httpx>=0.27.0

# Data Models
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# Frontend
streamlit>=1.38.0
